# perf: sqlparser faster formatting